% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_batch_mode.R
\name{dm_batch_mode}
\alias{dm_batch_mode}
\title{dm_batch_mode}
\usage{
dm_batch_mode(data_dir, results_dir, plots, pk_vl_save)
}
\arguments{
\item{data_dir}{directory holding multiple DLC .csv files for analysis}

\item{results_dir}{name of the directory that will hold the analysis results}

\item{plots}{TRUE to make plots, FALSE to suppress plots}

\item{pk_vl_save}{TRUE to create pk_vl_df.csv, FALSE to suppress}
}
\value{
parameters to start single file analysis
}
\description{
This is the starting script for batch mode analysis
}
