% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_DAQ_mode.R
\name{dm_DAQ_mode}
\alias{dm_DAQ_mode}
\title{dm_DAQ_mode}
\usage{
dm_DAQ_mode(plots, pk_vl_save)
}
\arguments{
\item{plots}{TRUE to make plots, FALSE to suppress plots}

\item{pk_vl_save}{TRUE to create pk_vl_df.csv, FALSE to suppress}
}
\value{
parameters to start kin DAQ analysis
}
\description{
This is the starting script for DAQ mode analysis
}
