% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_single_mode.R
\name{dm_single_mode}
\alias{dm_single_mode}
\title{dm_single_mode}
\usage{
dm_single_mode(data_dir, video.ID.no, plots, pk_vl_save)
}
\arguments{
\item{data_dir}{directory holding DLC .csv file for analysis}

\item{video.ID.no}{the video ID of the file to analyze}

\item{plots}{TRUE to make plots, FALSE to suppress plots}

\item{pk_vl_save}{TRUE to create pk_vl_df.csv, FALSE to suppress}
}
\value{
parameters to start single file analysis
}
\description{
This is the starting script for single file mode analysis
}
